Spliting Elastic Utils into three Utils file, for each index app. Needs test adjustments
