+++ elastic/utils.py
@@ -2,20 +2,18 @@ import logging
 import json
 import math
 import requests
-import pandas as pd
 import urllib3
 import homura
 import tempfile
+import pandas as pd
 
 from datetime import datetime
 from urllib import parse
-from json.decoder import JSONDecodeError
 from requests.adapters import HTTPAdapter
 
 from django.conf import settings
 from rest_framework import status
 
-from detection import serializers, models
 from elastic import models as elastic_models
 
 urllib3.disable_warnings()
@@ -23,7 +21,7 @@ urllib3.disable_warnings()
 logger = logging.getLogger('django')
 
 
-class UtilFunctions:
+class Utils:
     """Util Class for generic code on dealing with ES data."""
 
     now = datetime.now()  # Timing full process.
@@ -89,102 +87,6 @@ class UtilFunctions:
             logger.warning(log)
             raise ValueError(log)
 
-    def serialize_detection_file(self, json_file: object) -> pd.Series:
-        """Method for serializing the uploaded json object.
-
-        Serialize data into a Panda Series.
-
-        Args:
-            json_file (object): The json loaded from the json file uploaded
-
-        Raises:
-            ValueError: Raises error if there is any problem serializing data
-
-        Returns:
-           pd.Series : Returns a new Pandas Series to be dealt further
-        """
-        logger.info(f'[{datetime.now() - self.now}] serializing....')
-        try:
-            serializer = serializers.DetectionSerializer(
-                data=json_file['features'], many=True
-            )
-            serializer.is_valid(raise_exception=True)
-            logger.info(f'[{datetime.now() - self.now}] Serialized!')
-            return self._create_detection_series(serializer.validated_data)
-        except Exception as exc:
-            log = f'Internal error: {str(exc)}'
-            logger.warning(log)
-            raise ValueError(log)
-
-    def serialize_soy_file(self, text_array_file: list) -> pd.Series:
-        """Method for serializing a read array file into a panda Series.
-
-        It follows the same flow as detection, so when it gets more complex,
-        it just needs some adjustments.
-
-        Args:
-            text_array_file (list): The read file as an array,
-                each element as a line
-
-        Raises:
-            ValueError: Raises error if it encounters and error
-
-        Returns:
-            pd.Series: A Panda Series with all bulk data needed for ES
-        """
-        logger.info(f'[{datetime.now() - self.now}] serializing....')
-
-        try:
-            logger.info(f'[{datetime.now() - self.now}] Serialized!')
-            return self._create_soy_series(text_array_file)
-        except Exception as exc:
-            log = f'Internal error: {str(exc)}'
-            logger.warning(log)
-            raise ValueError(log)
-
-    def load_detection_file(self, file_url: str) -> object:
-        """Loads a uploaded json file into a json object.
-
-        Args:
-            file_url (str): url for file download
-
-        Raises:
-            ValueError: Raises error if its not possible to parse Json File
-
-        Returns:
-            object: Json Object parsed
-        """
-        try:
-            file_content = self._download_file_from_url(file_url, '.geojson')
-            return json.loads(file_content.read())
-        except JSONDecodeError:
-            log = f'Unexpected sent json data.'
-            logger.warning(log)
-            raise ValueError(log)
-        except Exception:
-            log = f'File not found.'
-            logger.warning(log)
-            raise ValueError(log)
-
-    def load_soy_file(self, file_url: str) -> list:
-        """Download soy file and loads it into memory.
-
-        Args:
-            file_url (str): Url string for downloading the file
-
-        Raises:
-            ValueError: Raises error if file is not found in the url
-
-        Returns:
-            object: Returns an list of all lines on the file
-        """
-        try:
-            text_file = self._download_file_from_url(file_url, '.txt')
-            return text_file.readlines()
-        except Exception:
-            log = f'File not found.'
-            logger.warning(log)
-            raise ValueError(log)
 
     def _download_file_from_url(self, file_url: str, extension: str) -> object:
         """Download file from a specific sent url.
@@ -204,41 +106,6 @@ class UtilFunctions:
             return json_file
         except Exception:
             raise
-
-    def _create_detection_series(self, data: object) -> pd.Series:
-        """Creates a Panda Series with a serialized data.
-
-        Args:
-            data (object): Serialized and validated Json data
-
-        Returns:
-            pd.Series: returns a Panda Series with all Detection Models
-        """
-        logger.info(f'[{datetime.now() - self.now}] creating series....')
-
-        pd_detections = pd.Series([models.Detection(**value) for value in data])
-
-        logger.info(f'[{datetime.now() - self.now}] series created!')
-        return pd_detections
-
-    def _create_soy_series(self, data: object) -> pd.Series:
-        """Internal method for creating a Panda Series of Soy data.
-
-        Args:
-            data (object): The list of all the lines in the soy file
-
-        Returns:
-            pd.Series: A panda Series with all the bulk data, serialized
-        """
-        logger.info(f'[{datetime.now() - self.now}] creating series....')
-
-        index_list = [value for key, value in enumerate(data) if not key % 2]
-        value_list = [value for key, value in enumerate(data) if key % 2]
-
-        pd_soy = pd.Series([f'{i}{v}' for i, v in zip(index_list, value_list)])
-
-        logger.info(f'[{datetime.now() - self.now}] series created!')
-        return pd_soy
 
     def send_bulk_list(
         self,
@@ -281,7 +148,7 @@ class UtilFunctions:
             )
 
             body = [
-                self._get_bulk_string(t, es_structure)
+                self._get_bulk_string(t)
                 for t in bulk_list[lower_limiter:higher_limiter]
             ]
 
@@ -318,21 +185,12 @@ class UtilFunctions:
         logger.info(f'[{datetime.now() - self.now}] Sent')
         return insertion_errors
 
-    def _get_bulk_string(self, element: object, es_structure) -> str:
-        """Internal method for getting the bulk data.
-
-        It separates between Soy and Detection since both file have
-        a different structure
-
-        Args:
-            element (object): the element to get its bulk
-            es_structure ([type]): type of Structure we're dealing with
-                (Soy or Detection)
+    def _get_bulk_string(self) -> str:
+        """Internal function to return, for each index, its bulk line
+        
+        It needs an definition for each index app created on the service
 
         Returns:
-            str: Returns the bulk line for that element
+            str: the bulk line into single line string
         """
-        if es_structure.identifier == 'Detection':
-            return element.get_es_insertion_line()
-
-        return element
+        pass
