+++ soy/views.py
@@ -4,7 +4,8 @@ from datetime import datetime
 from requests.models import Response
 from rest_framework import generics, response, status
 
-from elastic import models, utils
+from elastic import models
+from soy import utils
 
 logger = logging.getLogger('django')
 
@@ -17,7 +18,7 @@ class UpdateSoyView(generics.CreateAPIView):
     """
 
     def __init__(self):
-        self.util_class = utils.UtilFunctions()
+        self.util_class = utils.Utils()
 
     def post(self, request: object) -> Response:
         """Post method that does the full process.
@@ -31,7 +32,7 @@ class UpdateSoyView(generics.CreateAPIView):
             Response: Response object
         """
         try:
-            es_structure, _ = models.ElasticStructure.objects.get_or_create(
+            es_structure, _ = models.Structure.objects.get_or_create(
                 identifier='Soy'
             )
 
@@ -75,7 +76,7 @@ class ClearSoyStructiore(generics.DestroyAPIView):
     """Clear Soy Elastic Structure view"""
 
     def __init__(self):
-        self.util_class = utils.UtilFunctions()
+        self.util_class = utils.Utils()
 
     def delete(self, request: object) -> Response:
         """Delete method for dealing with ES Structure clearing
@@ -90,7 +91,7 @@ class ClearSoyStructiore(generics.DestroyAPIView):
             f'[{datetime.now() - self.util_class.now}]'
             f' Clearing structure...'
         )
-        es_structure = models.ElasticStructure.objects.get(identifier='Soy')
+        es_structure = models.Structure.objects.get(identifier='Soy')
 
         self.util_class.delete_es_structure(es_structure)
 
@@ -103,7 +104,7 @@ class CreateSoyStructure(generics.UpdateAPIView):
     """Create Soy Structure of ES Server view"""
 
     def __init__(self):
-        self.util_class = utils.UtilFunctions()
+        self.util_class = utils.Utils()
 
     def put(self, request: object) -> Response:
         """Put method for dealing with inserting a new ES Structure into
@@ -117,7 +118,7 @@ class CreateSoyStructure(generics.UpdateAPIView):
         logger.info(
             f'[{datetime.now() - self.util_class.now}]' f' Creating structure...'
         )
-        es_structure = models.ElasticStructure.objects.get(identifier='Soy')
+        es_structure = models.Structure.objects.get(identifier='Soy')
 
         self.util_class.create_es_structure(es_structure)
